Adding a read_enabled bit.
Rui suggested that we need to have a read enabled bit -- it doesn't make
entire sense since we have a 'start' signal, but knowing that this circuit
will be driven by a clock, it makes sense. E.g. the 'start' signal will
always be alternating, so we use the read/write enabled to let our
internal's know what should be happening.

This also allows us to perform
read and write in the same cycle... This might be bad behavior...