Updating testbench, added note.
The testbench is being updated to have less duplicate code.

Also added a note to the register_file implementaiton since I noticed that
the finished bit is not following the rules of combinational circuits,
essentially the circuit is finishing at the same time as it is starting so
the finish bit is never getting a 'low' ripple. Unsure how to get the low
ripple without faking via wire delays.