Finished adding and testing JR instruction
There is some really weirdness because we assume that the incoming
bitstream will have all zeros for the rd and rt registers

e.g. we assume the instruction will look like
0000 00ss sss0 0000 0000 0000 0000 1000

This concept was taken from:
http://www.mrc.uidaho.edu/mrc/people/jff/digital/MIPSir.html

This does mean that our control unit now takes the funct as input since
determining whether or not an r-type instruction is a JR instruction, we
need to read the funct. This should be easily available from the Decoder
module when composing these modules.