Both Memory and RF write on negedge. RF combo read
Changed both memory and register file to perform their writing in the
negatrive edge, meaning that while clock is high, we perform all reads and
combinational logic in our CPU, then during the time when clock edge is
low (negative) is when we do all of our writing back.

This will work since StoreWord is mutually exclusive from all of the other
register file writing functions.

I.e. we will never perform a write to memory in the same cycle as a write
to registerfile.