Got running on board using Key -- details below
Got running using Key[1]
Reset works, instructions execute.
Made it so that instruction memory is the right size (was off by one)
Made it so that the instruction loader loads a jump to Exit at the end of
the program.
Made ALU not use 'don't care' but instead 0 in the default case